Extract results graph to separate file
