+++ src/visitPage.js
@@ -1,4 +1,3 @@
-import _ from 'lodash'
 import React, { Component } from 'react';
 import { Grid, Step } from 'semantic-ui-react';
 import { Redirect, Prompt } from 'react-router-dom';
@@ -17,7 +16,7 @@ import { clientList } from './clientList';
 // Our Components
 import SimpleMenu from './simpleMenu';
 import AlertSidebar from './alertSidebar'
-import { FormPartsContainer } from './forms/formHelpers';
+import ResultsGraph from './resultsGraph';
 import { PreviousIncomeStep } from './forms/previousIncome';
 import { CurrentIncomeStep } from './forms/currentIncome';
 import { PreviousExpensesStep } from './forms/previousExpenses';
@@ -36,158 +35,6 @@ const StepBar = ({ steps, currentStep }) => {
 
   return (<Step.Group size='mini' ordered items={steps} />)
 }
-
-const Results = (props) => {
-  var xRange = _.range(0, 100000, 1000);
-  /** Need a new object so client's data doesn't get changed. */
-  var fakeClient = { ...props.pageState };
-
-  var massHealthData = xRange.map(x => {
-      fakeClient.annualIncome = x;
-      return getMassHealthEligibility(fakeClient).benefitValue});
-    
-  var snapData = xRange.map(x => {
-      fakeClient.annualIncome = x;
-      return getSnapEligibility(fakeClient).benefitValue});
-
-  /** Section-8 Housing Choice Voucher */
-  /** @todo Base this rent on FMR areas and client area of residence if no rent available. */
-  fakeClient.previousContractRentMonthly = 700;
-  fakeClient.previousEarnedIncomeMonthly = 0;
-  var housingData = xRange.map(function ( annualIncome ) {
-    // New renting data
-    var oldRentShare = fakeClient[ 'previousRentShareMonthly' ];
-    fakeClient.currentEarnedIncomeMonthly = annualIncome/12;
-
-    var result  = getHousingBenefit(fakeClient),
-        subsidy = result.benefitValue * 12;
-
-    // Prep for next loop
-    var newShare = result.data.newRentShare
-    fakeClient[ 'previousRentShareMonthly' ] = newShare;
-    fakeClient.previousEarnedIncomeMonthly   = annualIncome/12;
-
-    return subsidy;
-  });
-
-  var data = {
-    labels: xRange,
-    datasets: [{
-        label: "MassHealth",
-        borderColor: "rgba(206, 125, 61, 1)",
-        data: massHealthData,
-        fill: false,
-    },
-    {
-      label: "SNAP",
-      borderColor: "rgba(101, 47, 138, 1)",
-      data: snapData, //xRange.map(x => ({ annualIncome: x, householdSize: props.pageState.householdSize }).benefitValue),
-      fill: false
-    },
-    {
-      label: "Section 8 Housing",
-      borderColor: "rgba(206, 203, 61, 1)",
-      data: housingData, //xRange.map(x => getHousingBenefit({ annualIncome: x, householdSize: props.pageState.householdSize }).benefitValue),
-      fill: false
-    },
-    ]};
-
-  var options = {
-    title: {
-      display: true,
-        text: 'Benefit Eligibility for Household Size ' + 
-                props.pageState.householdSize
-    },
-    showLines: true,
-    scales: {
-        yAxes: [{
-          scaleLabel: {
-            display: true,
-              labelString: 'Benefit Value ($)'
-          },
-          ticks: {
-              beginAtZero: true,
-              /*
-               * function to add $ and 1,000s separators to graph axes
-               * we are using chart.js v2.7 so it requires a callback function
-               */
-              callback: function(label) {
-                  return label.toLocaleString("en-US");
-              }
-          }
-        }],
-        xAxes: [{
-          scaleLabel: {
-            display: true,
-              labelString: 'Annual Income ($)'
-          },
-            ticks: {
-              callback: function(label) {
-                  return label.toLocaleString("en-US");
-              }
-            }
-        }]
-    },
-      /*        
-       * default tooltip for chart.js 2.0+  when unspecified looks like:
-       *
-       * options: {
-       *   tooltips: {
-       *       callbacks: {
-       *           label: function(tooltipItem, data) {
-       *               return tooltipItem.yLabel;
-       *           }
-       *       }
-       *   }
-       * }
-       *
-       */
-    tooltips: {
-        callbacks: {
-            // format the title of the tooltips to be in USD
-            title: function(tooltipItems, data) {
-                return data.labels[tooltipItems[0].index].toLocaleString("en-US",
-                    {style:"currency",currency:"USD"}).replace('.00','');
-            },
-            /*
-             * to add number formatting to the tooltips. returns the data label 
-             * + currency format 
-             * from https://github.com/chartjs/Chart.js/issues/2386
-             */
-            label: function(tooltipItem, data) {
-                return data.datasets[tooltipItem.datasetIndex].label + ": " + 
-                    tooltipItem.yLabel.toLocaleString("en-US",{style:"currency", 
-                        currency:"USD"}).replace('.00','');
-            }
-        }
-    }
-  };
-
-  // return (
-  //   <wrapper className = 'result-page'>
-  //     <FormPartsContainer
-  //       title     = {'Results'}
-  //       left      = {{ name: 'Go Back', func: props.previousStep }}
-  //       right     = {{ name: 'Save Results', func: () => props.saveForm(false) }}
-  //        <div> <Line data={data} options={options} /> </div>
-  //     </FormPartsContainer>
-  //   </wrapper>
-  // )
-
-  // Non-saving version for first prototype testing
-  return (
-    <wrapper className = 'result-page'>
-      <FormPartsContainer
-        title     = {'Results'}
-        left      = {{ name: 'Go Back', func: props.previousStep }}
-      >
-         <div> <Line data={data} options={options} /> </div>
-      </FormPartsContainer>
-    </wrapper>
-  )
-
-};  // End Results()
-
 
 class VisitPage extends Component {
   constructor(props) {
@@ -238,7 +85,7 @@ class VisitPage extends Component {
       { completed: false, active: false, title: 'MassHealth', form: HealthStep },
       // { completed: false, active: false, title: 'SNAP', form: SNAPStep },
       // { completed: false, active: false, title: 'Housing', form: HousingStep },
-      { completed: false, active: false, title: 'Results', form: Results }
+      { completed: false, active: false, title: 'Results', form: ResultsGraph }
     ];  // end this.steps {}
 
     this.stepProps = {
